Update payload.txt
This is perfectly simple payload and I think it's great! I did have trouble with the second "exit" until I increased the delay. Then it worked fine! Was trying to get it to output the log.txt using the computer's get-computerinfo -p CsName in the file name, but just couldn't get the syntax correct. Perhaps you have a suggestion?